Update CCI RPOP IR1101.ftl
resolved a number of issues and was able to get this to load successfully on IR1101 without error and still remain online with rainier. previous version killed rainier connection. also added if clauses where appropriate. next step is to try to connect to a real VPN site to site. also removed reference to gig 0/5/0 used for IOX, not clear if IOX is needed for CCI.