Update setAPvlan script and resiliency config
Changed the native vlan set in the EEM script for WGB config to 50 and added "no spanning-tree vlan 50" to ensure the AP is reachable and functional. Also wrapped the statement that removes the default route added by the bootstrap template in an if statement checking if the first cellular interface is enabled.